enlarge [min, max] range of GlQuantitative integer scale (use long long instead of int) + Histogram bug fix: ensure [min, max] range of double values can fit in [min, max] range of long long values before using an integer scale on an Histogram X axis
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@10073 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ plugins/view/ParallelCoordinatesView/QuantitativeParallelAxis.cpp
@@ -73,9 +73,9 @@ void QuantitativeParallelAxis::setAxisLabels() {
     integerScale = false;
   }
   else {
-    int min = static_cast<int>(axisMinValue);
-    int max = static_cast<int>(axisMaxValue);
-    int incrementStep = (max - min) / DEFAULT_NB_AXIS_GRAD;
+    long long min = static_cast<long long>(axisMinValue);
+    long long max = static_cast<long long>(axisMaxValue);
+    unsigned long long incrementStep = (max - min) / DEFAULT_NB_AXIS_GRAD;
 
     if (incrementStep < 1) incrementStep = 1;
